fix: align perps chart and footer behavior
