+++ packages/kit-bg/src/states/jotai/atoms/perps.ts
@@ -1097,7 +1097,7 @@ export const {
     skipOrderConfirm: false,
     showTradeMarks: true,
     showChartLines: true,
-    chartPosition: 'top',
+    chartPosition: 'bottom',
     hideSmallSpotHoldings: true,
     lastTriggerOrderType: ETriggerOrderType.TRIGGER_MARKET,
     lastAdvancedOrderType: ETriggerOrderType.TRIGGER_MARKET,
